requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.25
Werkzeug==2.3.7
flask-login==0.6.2
python-dotenv==1.0.0
Flask-CORS==4.0.0
ntplib==0.4.0
